+++ src/Util/PropertyValueUtil.php
@@ -13,20 +13,13 @@ namespace Eluceo\iCal\Util;
 
 class PropertyValueUtil
 {
-    public static function escapeValue($value)
-    {
-        $value = self::escapeValueAllowNewLine($value);
-        $value = str_replace("\n", '\\n', $value);
-
-        return $value;
-    }
-
-    public static function escapeValueAllowNewLine($value)
+    public static function escapeValue(string $value): string
     {
         $value = str_replace('\\', '\\\\', $value);
         $value = str_replace('"', '\\"', $value);
         $value = str_replace(',', '\\,', $value);
         $value = str_replace(';', '\\;', $value);
+        $value = str_replace("\n", '\\n', $value);
         $value = str_replace([
             "\x00", "\x01", "\x02", "\x03", "\x04", "\x05", "\x06", "\x07",
             "\x08", "\x09", /* \n*/ "\x0B", "\x0C", "\x0D", "\x0E", "\x0F",