+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 ### Changed
 - Drop support for old PHP versions: 5.3, 5.4, 5.6
 
+### Removed
+- Remove class `\Eluceo\iCal\Property\Event\Description` (see [#61](https://github.com/markuspoerschke/iCal/pull/61))
+- Remove static method `\Eluceo\iCal\Util\PropertyValueUtil::escapeValueAllowNewLine` (see [#61](https://github.com/markuspoerschke/iCal/pull/61))
+
 ## [0.11.1] - 2017-04-04
 ### Fixed
 - All days events (no time) ends on the next day. [#83](https://github.com/markuspoerschke/iCal/pull/83)